pyproject.toml
[project]
name = "mlx8-w4-multimodal-transferlearning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.6",
    "pyarrow",
    "torchvision",
    "datasets",
    "tqdm",
    "wandb",
    "colorama",
    "pandas",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "numpy<2",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "scikit-learn>=1.7.0",
    "hdbscan",
    "scipy",
    "transformers",
    "qwen-vl-utils[decord]==0.0.8",
    "accelerate",
    "torchaudio",
    "bitsandbytes",
    "unsloth-zoo",
    "litellm>=1.73.6",
    "duckdb>=1.3.1",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "unsloth>=2025.6.12",
]

#   "unsloth[cu121-ampere-torch230]",
 
# uv add "unsloth[cu121-ampere-torch231] @ git+https://github.com/unslothai/unsloth.git"
# [tool.uv.sources]
# unsloth = { git = "https://github.com/unslothai/unsloth.git" }

# [tool.uv]
# no-build-isolation-package = ['flash-attn', 'setuptools']